Check index.docker.io/v1 when auth.docker.io/token has no auth
The Docker daemon will check an older auth store value for challenges,
so we have to as well.

Fixed a bug where image stream import wasn't setting spec tags (which
meant oc import-image couldn't be invoked again if the first time
failed) and where the conditions weren't getting cleared if a second
import ran successfully with the same output image.