Resource quota for images and image streams
New resource quotas:

1. openshift.io/projectimagessize
2. openshift.io/imagestreamsize
3. openshift.io/imagesize

The 1st allows to limit a total size of images across all repositories
in a namespace.

The 2nd allows to set a limit for a sum of sizes of all image layers in
a single image stream (aka repository). Usage will always be 0 because
we cannot store different values for different repositories.

The 3rd allows to set a limit for a sum of sizes of all layers of
single image. Usage (soft-limit) will always be 0.

Layers occuring multiple times in a single repository count just once.

Signed-off-by: Michal Minar <miminar@redhat.com>