Force dind deployment to build binaries by default
A recent change to hack/build-go.sh ensured that binaries were sourced
from a release tar if present.  This change allows that behavior to be
overridden with OS_RELEASE=n so that dind can set this by default.  This
is intended to allow ci to validate that dind can be deployed without
error from a clean environment.