Ensure NodeIP doesn't overlap with the cluster network
If a HostSubnet is edited or the node's IP address is changed such
that the NodeIP is contained within its own SubnetCIDR, a routing loop
will be created in the OVS bridge and cause a kernel panic by
exhausing kernel stack space.

For the master: warn about existing invalid subnets at startup, warn
if they are created when the cluster is running, and refuse to create
a HostSubnet resource for an invalid node IP address.  But don't exit
since that prevents fixing the problem with 'oc edit hostsubnet'.

For the node: exit if the node's HostSubnet contains an invalid IP,
and ignore other nodes' invalid HostSubnets so that we don't add flow
rules to them.

https://bugzilla.redhat.com/show_bug.cgi?id=1295486