router: use explicit maxUnavailable number of pods
The way to calculate maximum unavailable pods for a deployment recently
changed to use math.Floor (over math.Ceil it was using before), meaning
for a deployment with size N=1, maxUnavailable=25% previously would
resolve to 1, now it resolves to 0. This commit switches router's
maxUnavailable to use an explicit number instead of a percent. 25%
doesn't make sense for less than 4 pods and takes effect only after 8
instances of a router are up and running.