Flexible dockerfile builder that uses client calls
This commit adds a library for executing Dockerfile builds without a
direct dependency on the Docker daemon, and is able to create an entire
Dockerfile into a squashed layer directly. It uses the upload and exec
calls for COPY, ADD, and RUN, and reuses the same logic as the existing
builder for transforming environment variables and arguments.