Allow multiple web login methods
Retain the order of multiple login mechanisms

Save an array of login mechanisms that corresponds to the order
that each identity provider appears in the master-config.yaml file
so that the order the user sees is constant.

Don't use IDP-specific path for sole password provider

If there is exactly one password provider in the configuration,
allow it to use the simple /login path rather than /login/<IDP name>

Make openShiftLoginPrefix private

When multiple password-based IDPs are present, this prefix would
not be accurate. We will not expose this value publicly so that
other modules will not attempt to rely upon it.