Include dockerImageManifest and dockerImageConfig in the ImageStreamImage
This information is useful for users and not just admins.
Interfaces use this info to render details about the layers that
an image contains. Until this information is in another form it
should be consistently visible with the image.

Obviously it would be nice if this wasn't escaped, but that seems
difficult to accomplish with golang's JSON decoding.