F5: handle HTTP 409 responses gracefully
Because of an API change between F5 BIG-IP 11.5 and 11.6, the existence
checks for certain resources can return false negatives, in which case we
will get HTTP 409 error responses when we try to create those resources.
Instead of treating such responses as errors, log them, and then continue
as if we had created the resources successfully.

Make this change for resources that we create on initialization: policies,
associations between vservers and policies, iRules, and datagroups, as well
as pools and pool members.  The reasoning is that these resources have
distinctive, descriptive names, and it is improbable that a different
resource will exist with the same name.  On the other hand, policy rules
and rule conditions and actions have non-descriptive numeric names, and
keys and certificates are named based on the route name (not the key or
certificate data), so it is more probable that an existing resource of any
of these types will be different from the resource we want to exist.  Thus
we still treat HTTP 409 responses as errors for these resources.

This commit fixes https://github.com/openshift/origin/issues/7566