Fixes the ip address to namespace cache when pods reuse addresses
There's a cache that tracks which namespace IP addresses belong to.
However, the cache logic was not considering that Failed or Complete
pods (i.e. not running) could share an IP address with a Running pod.
So the cache code needed to be changed to only add items to the cache
when a pod was Running or Pending.  We also track the pod UID to make
sure that when we delete, we are deleting the entry for the pod that
was running.  So that if a stopped pod is edited, that entry doesn't
end up in the cache.

This is reproducible by setting the networkConfig hostSubnetLength to
3 in the master and node config on your cluster.  Then create two
namespaces, ns1 and ns2, a service in ns2 that has a selector that
matches both pods then create a pod that fails in one (have it run
"sleep 5; false", and set the restartPolicy to Never.  Once it has
failed, create a pod that runs correctly in the other (hello-openshift
is fine).  Then curling the service URL should work.  Remove the
failed pod and the curl will then fail.

The problem is that the cache sees the failed pod removal and removes
the ip from the ip->namespace cache.  But it should not have been
removed because the running pod put the entry in.

For bug:
  https://bugzilla.redhat.com/show_bug.cgi?id=1322942

Also-Authored-By: sross@redhat.com