debug.sh: fix incoming remote test packet traces
The current OpenFlow rules will drop remote packets that don't have a
valid tun_src, meaning we have to specify that now when generating
test packets to trace.

Also, we were specifying the VNID to test as reg0, but this is wrong
for local pods (since we should be letting the flow rules fill in reg0
correctly themselves) and wrong for remote pods (since reg0 gets
overwritten from tun_id, which we weren't setting).