Add a mutex to registry.podsByIP
The pod-monitoring and endpoint-filtering threads might both run at
the same time so we need a mutex here.

As for the actual race condition, if a pod which is an endpoint is
added, and OnEndpointsUpdate() ends up running before trackPod(), then
the pod won't be allowed as an endpoint and will be ignored until the
next time OnEndpointsUpdate() runs, when it will get picked up. So
isolation is guaranteed but correct functioning of the service is
not...