add default resource requests to router creation
This is a repurposing of my original pull request to add a default
resource request to router creation instead of settable resource
requests and limits.

100 millicores was an agreed upon default for CPU and I observed that
the basic router with no routes uses about 86MB of RAM and decided 256MB
was a good starting point for a request that allows ample room for
growth.