Refactored stacktrace implementation for bash scripts
This changes the prior implementation of the stacktrace to use the
new stacktrace implementation as well as the newly-implemented Bash
trap framework for the ERR signal.

In order to make the stacktrace output more useful, especially for
someone attempting to re-run the commands used, we evaluate all of
the variables in the $BASH_COMMAND as best we can. We will not be
able to ever correctly evaluate positional variables ($1, $2, $@)
or variables created in the local scopes of the trap handler or in
the stacktrace function such as $return_code, $last_command, or
$errexit_set. For this reason, we print both $BASH_COMMAND and our
attempt to substitute variables.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>