tests: Bind DNS only to API_HOST address
Tests may fail on many developer systems because the default
configuration attempts to bind port 53 on all interfaces. This means
that systems running libvirt or dnsmasq for local caching will fail
and be unable to run the tests. With this patch, we restrict the DNS
to bind only to the public address of the API_HOST.

Signed-off-by: Stephen Gallagher <sgallagh@redhat.com>