Prevent deployer pod creation conflicts
Take advantage of a single-controller assumption by checking for the
existence of deployer pods before trying to create new ones. This allows
the controller to progress the phase of the deployment without incurring
a resource conflict during create under most circumstances- a behavior
which can result in poor interactions with the quota system which will
charge quota for the conflicting pod.