haproxy Cookie id leaks information about software
Resolves: bz1318796
haproxy now uses environment variables, set in the deploymentconfigs
to set the cookie prefix and sufix. When the environment variables are
not set, the original cookie name is generated which prevents breaking
software that relies on the structure of the cookie name.

There are three environment variables to set the prefix on the three
distince cookies that can be generated. When a prefix is set the
"project_routename" is hashed to obfuscate it. The suffix environment
variable sets the suffix.
COOKIE_PREFIX_MAIN       default "OPENSHIFT_"
COOKIE_PREFIX_EDGE       default "OPENSHIFT_EDGE_"
COOKIE_PREFIX_REENCRYPT  default "OPENSHIFT_REENCRYPT_"
COOKIE_SUFFIX            default "_SERVERID"

Signed-off-by: Phil Cameron <pcameron@redhat.com>