Allow network test runner to use kubeconfig
Previously the networking extended test runner required that the root
of a config tree be specified to target an arbitrary cluster.  This
change adds support for the configuration file to be specified via
OPENSHIFT_TEST_KUBECONFIG.  KUBECONFIG is not used due to jenkins
having this set in the slave environment to an invalid value.