Refactored binary build shell invocation
By changing the amount of code that is literally contained
in the sub-shell `()` braces, we allow the stacktrace that
is generated to contain much less text on any one line, so
that it is read-able and useful. Furthermore, by naming the
positional variables we pass around we allow the new stack
trace code to correctly evaluate them and show us what went
wrong, exactly.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>