Refactored command functions not to accept a Writer
If a consumer of the command wants to write the standard output from
the command to another source, they can do so manually after the func
returns the full command output. There was only one consumer of this
option and it made things significantly more complicated, so it seems
more reasonable to factor this way.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>