Project should not fetch client immediately
oc project can be used to switch config contexts, but loads its client
in the Complete() method which does API negotiation. Since a user may
not be logged in or the server they are talking to may be down, changing
projects then doesn't work.

Use a lazy client loader just for project until we can sort out lazy api
negotiation and/or something better on client.