Cache limit range objects in registry
To speed up resource usage checks during blob upload in the registry,
a cache is added to hold limit range objects.

They are stored in expiring cache with a configurable TTL.

The TTL is a number of seconds the cache objects are retrievable from
the cache. Its value can be specified with
REGISTRY_MIDDLEWARE_REPOSITORY_OPENSHIFT_PROJECTCACHETTL environment
variable.

Signed-off-by: Michal Minar <miminar@redhat.com>