router: add --override-exceptions flag
Add --force-exceptions flag for `oadm router` and --override-exceptions
flag and ROUTER_OVERRIDE_EXCEPTIONS environment variable for the router
itself.

Add ForceExceptions string array field to RouterConfig and update
NewCmdRouter and RunCmdRouter accordingly.

Add OverrideExceptions string array field to RouterSelection and update the
RouterSelection Bind and Complete methods accordingly.

Add the exceptions parameter to NewRouteValidator and routeSelectionFunc
(the former passing it to the latter).

Update callers of NewRouteValidator to pass in the exceptions parameter.

Modify routeSelectionFunc to check the exceptions before overriding
a route's host and leave it as is if the route's namespace is in the list
of exceptions.

Add TestOverrideExceptions.

Regenerate completions.