router: add --allow-custom-certificates flag
Add an --allow-custom-certificates flag (default true) for `oadm router`
and for the router, and ROUTE_ALLOW_CUSTOM_CERTIFICATES environment
variable for the router, that specify whether routes that specify custom
certificates will be accepted.

Namespaces specified with --force-exceptions/--override-exceptions will
still be allowed to specify custom certificates even if
--allow-custom-certificates=false is specified.

Add AllowCustomCertificates Boolean to RouterConfig and update NewCmdRouter
and RunCmdRouter accordingly.

Add AllowCustomCertificates Boolean to RouterSelection and update the
RouterSelection Bind and Complete methods accordingly.

Add allowCustomCertificates and overrideExceptions to RouteValidator.

Add a new parameter for the flag to NewRouteValidator, and update
NewRouteValidator to set the new fields in RouteValidator.

Update RouteValidator's HandleRoute method to reject any route with
a custom certificates if the route's namespace is not in the list of
exceptions and the new flag is set.

Update callers of NewRouteValidator to pass in the new flag.

Add TestAllowCustomCertificates.

Regenerate completions.