sdn: use CNI for IPAM instead of docker
Leave docker configuration alone.  Instead, let docker do whatever it
wants to do, then pull the interface out of the docker bridge, clear
the docker assigned addresses and routes, and allocate an address
with the CNI 'host-local' plugin from the node subnet.

Note that this commit breaks (a) HostPort functionality (which docker
used to manage since it handled IPAM) but that is fixed in the next
commit, and (b) a docker-added iptables masquerade rule that is fixed
in a later commit as well.