Support a loopback listener for in process communication
This uses chanstream (a fork of github.com/gdamore/chanstream) to allow
the all-in-one to connect to itself (and etcd) over in process memory
channels rather than network connections. etcd and the master "listen"
on a named string equal to their public address, and then clients use a
custom dialer to try connecting to that public address via loopback
rather than opening a network connection.

The hardest part is putting the loopback listener before the TLS
listener is initialized, since etcd does a poor job of being
"wrappable", as do a few other servers.

Needs some more testing and experimentation to figure out whether this
is worthwhile. Avoids a lot of syscalls.