Updated spdysteream logger, forced flake in e2e
The spdystream logger was being clobbered by k8s custom
logging solutions, so in order to see the output we had
to simply put it into `fmt.Printf`.

In order to force the e2e `oc run --attach` flake to
manifest in every Jenkins test job, it is placed in a
while loop.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>