Elongate pod lifetimes for pods used in `oc run --attach`
Wedges in the `oc run --attach` call were previously
caused most likely by a race between the container
exiting and the attach mechanism attaching to the
file descriptors from the container. By keeping the
containers alive for a long time before they become
active, we should alleviate this flake.

The go-dockerclient library has a mechanism to
coordinate the pod creation / attach routines, but
Kubernetes has migrated away from the library, so it
makes little sense to refactor code in the kubelet
to use the library if the patch will be dropped in
the upcoming rebase.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>