Detect data races in `go test` log output
The `go test` race detector does not fail a unit test
if it finds a data race, and also does not print its
output to stderr, so we cannot present the findings
easily. Instead, we can detect that a race was found
and point to the logfile for more information.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>