Make node to auto detect openshift network plugin from master
Node config nodeConfig.NetworkPluginName is optional for openshift network plugin.
If plugin is not specified, it will auto detect by looking at ClusterNetwork.
If plugin is specified, we will validate both node and master are using same network plugin or not.
nodeConfig.NetworkPluginName is not deprecated to support third party network plugins.
We can't expect or enforce third party node network plugin to auto detect from master.

When openshift network plugin is used, fix check to block node start when kubelet is enabled but not plugin and viceversa.