Limit the amount of output of a faulty new-build
Old output of a mistyped or not found image in a new-build call
would return all possibly matching images.  This commit limits
that to five, and instead suggests using the new-app search function.
Does not make search suggestion if there are less than 5 matches.

As a side fix, we also now check for version prefixing, so v1.1
will partially match to v1.1.1 and v1.1.2.