dind: avoid being targeted by oci-systemd-hooks
oci-systemd-hooks detects that a container will run systemd by checking
for an entrypoint command of 'systemd' or 'init' and then mounts most
cgroups read-only to support running systemd in an unprivileged
container.  This is not compatible with dind containers, which require
running with --privileged to ensure cgroups mounted read-write.  This
change hardlinks 'init' to 'dind_init' and uses that as the entrypoint
to avoid having dind containers be targeted by the hook.