Add local pod watch and CNI setup block for hostport handling
Now that Docker is no longer doing network setup, something needs to handle
HostPort/ContainerPort mappings.  Use the upstream HostportHandler to do this.
However, it must have knowledge of all pods running on the node and keeps
hostports open for all pods (thus arbitrating conflicts) so it must run in
the openshift-node process rather than the CNI plugin which deals only with
one pod.

The HostPort code needs the pod's IP adddress, but this is not available
until after the CNI plugin has returned.  Returning from the CNI plugin
before hostports are set up could cause traffic to start flowing to the
HostPort before the mapping is completed.

To prevent this race:

1) the CNI plugin pushes the PodIP to the api-server
2) the CNI plugin then waits for /var/run/openshift-sdn/hostports/<UID>
to contain 'ok' or an error message
3) openshift-node waits for the pod's IP address to appear
4) openshift-node creates any hostport mappings required
5) openshift-node writes the status file to /var/run/openshift-sdn/hostports/<UID>
6) the CNI plugin returns success if it sees 'ok' or returns the error
from the status file
7) the CNI plugin deletes the status file