Use anonyous FIFO instead of pipe in os::cmd output test
Previously, the output of the last command was searched by
`grep -q`. When `grep -q` finds the text it is looking for,
it exits immediately with exit code 0. If the command feed-
ing `grep -q` with input is still running, it will write
to a pipe that has no readers, and Bash will terminate the
producer process with a SIGPIPE as this is invalid behavior.

Instead of using a `cat | grep` approach for searching in
os::cmd, we can therefore use `grep <(file)` approach, which
creates an anonymous FIFO to pass the data and therefore
guards against this sort of error.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>