Removed spam from hack/test-cmd.sh output on test success
Previously, something like 300-500 lines of spam were put
into stdout from the pprof output and the metrics data.
This data was very rarely useful and always mirrored in
logfiles on disk anyway, so this commit replaces those
spam messages with messages that point out the location
of the logfiles instead.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>