Migrated tests for login from hack/test-cmd.sh preamble
These tests remained in the hack/test-cmd.sh preamble for historical
reasons, but do not need to be there. The side effects that these
tests had in setting up the KUBECONFIG and KUBERNETES_MASTER envars
for the remaining tests are simply replicated in the preamble and
removed in the test to ensure that the tests are running the same
way that they were originally.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>