Handled working dir changes in Bash relative path util
The os::util::repository_relative_path utility attempts to print
the relative path from the repo root to a file in the repo. To
accomplish this, it utilizes `realpath` and is therefore dependent
on the working directory to be able to resolve relative paths. If
a script changes the current working directory, the utility breaks.
With this change, the working directory at the time of sourcing
the utility is recorded and used to make the function is robust to
that behavior.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>