Removed namespace assumptions from test-cmd test cases
A number of test-cmd test cases were making assumptions
about the namespace that they would be executing in. As
we support running the the tests outside of the script
launcher in hack/test-cmd.sh, tests cannot assume their
namespaces. This commit refactors the tests to determine
their namespace with `oc project -q` and thereby make
the tests less fragile.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>