Refactored the dependency between images and images-old-policy
In order to better support the manner in which both the images.sh
and images-old-policy.sh test-cmd test buckets are attempting to
run their tests, the tests have been factored out and are run in
an identical manner from each wrapper. The wrappers should choose
to alter the default jUnit test suite name so as to not alias the
test suites when they are compiled into the XML report.

Other changes in this commit include:
 - an image registry is started in hack/test-cmd.sh always
 - the diagnostics tests are updated to expect the registry to
   exist but to not be backed by any pods
 - Bash idioms and quoting are improved

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>