Stop using node selector as ipfailover label
The ipfailover dc and its template were having the label(s) provided
for the node selector set as their label(s).  If, as in the
documentation, the router service was selecting on that label, the
service would end up targeting the keepalived pods.  This change
ensures the dc and template define a new label rather than the one
provided as the node selector.