The first letter should be capitalized
Replace the tab with 4 spaces

Spaces, it is aligned to view code with any editor, including the web
page view (such as in the GitHub code). When use tabs, the view on the
page is chaos, and it is terrible to mix tabs and spaces.

Revert "Replace the tab with 4 spaces"

This reverts commit 54350ece13ad78662a4b60bd1cf6f79d53493179.