sdn: clear kubelet-created initial NetworkUnavailable condition on GCE
When running with the GCE cloud provider, kubelet code adds an initial
NetworkUnavailable condition to each node which is supposed to be cleared
by the cloud route controller.  Overlay networks (like flannel or
openshift-sdn) don't need the route controller for inter-node
communication, but kubelet code that adds the initial condition
only knows about flannel.

Until we can teach upstream kubelet code that some network plugins
don't need the initial NetworkUnavailable condition set, remove that
condition on the master when the node's status updates. This is simpler
than removing it on the node itself, because openshift-sdn and kubelet
startup are asynchronous, and the master already has a node watch.

https://bugzilla.redhat.com/show_bug.cgi?id=1357474