Allow users to choose database images version in the DB templates
Currently, MongoDB, MySQL and PostgreSQL templates are stuck with
a fixed version of the database images. Users cannot change the
version directly without exporting the templates and doing manual
editing and importing the templates back again.

A new environment variable <DB-NAME>_VERSION is added to each template
to allow to users to change it directly on UI or via CLI directly
before the submission. No more manual editing for DB templates.

Signed-off-by: Vu Dinh <vdinh@redhat.com>