Scale helpers: always fetch scale object for DCs
The scale helper for DeploymentConfigs previously would fetch the raw
deployment, covert that into a scale object, and then submit that scale
object as an update to the scale subresource of the DeploymentConfig.
This caused issues when using recent clients against older API servers
since new metadata (the UID) was added to scale, and the older API
servers would choke while validating the update.

Fixes bug 1370058