Fix SDN network plugin detection
When OpenShift master and node are started at the same time.
We could end up in a state where node registration succeeds but
API call to ClusterNetwork resource fails. We ignore the error
assuming that it's not openshift specific network plugin but this
can lead to master running openshift sdn plugin and node without
sdn plugin.
So the fix is to only ignore the error if ClusterNetwork API call
fails with 'not found' error. Anything else(api not ready, etcd error,
etc.) should throw an error.