Make the router data structures remove duplicates cleanly
The bug is caused by the modification of the route.  In the code it
would add a new object to the list of routes by the host and not
remove them.  When the route was deleted, it would remove all of the
routes with the same name, and then the list would be cleared.
Unfortunately, the code would then look at the number of items in the
list, and either delete it from the map, or clean items out of the
list.  But since it was possible for there to be multiple duplicate
names, we could end up with an empty list.  Later code would assume
that if there was a key in the map, then the associated list would
have items.  That caused a panic.

This fix changes both places:
- If there are duplicates, remove them all, and then if the list is
  empty, delete the key from the map
- When a modification happens, remove the previous entry from the list

So, the first duplicate check is now just for protection, but the
proper fix is to clean out the lists before adding items to them.

Fixes bug 1371826